requirement.txt
watchdog
websocket-client
zmq